requirements.txt
# Flask and extensions
Flask>=2.2.0
Flask-SQLAlchemy>=3.0.0
Flask-SocketIO>=5.3.0
Flask-Cors>=3.0.0
gunicorn>=20.1.0 # For production
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0

# Playwright for automation
playwright>=1.30.0

# Utilities
psutil>=5.9.0
pyautogui>=0.9.0
pygetwindow>=0.0.9 # Optional, for Windows
pywin32>=305 # Optional, for Windows

# For schema diagram generation
sqlalchemy-schemadisplay>=1.3
Pillow>=9.0.0
pydot>=1.4.2 # Requires Graphviz installation

# Testing
pytest>=7.0.0